Split guess service into authed and unauthed flow.
While this created duplications in a few places and hid some of the
implementation details in the superclass, I believe it reflects a more
of a simple green approach. That is, both the duplication and variance
becomes more obvious and easier to refactor down the line.

I kept an unauthed flow here rather than require users because I think
that is more typical of problems I've actually seen on projects. Those
requirements are exactly where the complexity comes from and I really
didn't want to shy away from that. I mean that's why this stuff is hard.